seperated business logic
