+++ app/(pages)/wallet.tsx
@@ -1,67 +1,27 @@
 import { View, Text } from 'react-native'
 import NavButton from '../../components/NavButton'
 import { useEffect, useState } from 'react';
-import { Buffer } from 'buffer';
-import EC from 'elliptic-expo/lib/elliptic/ec';
-import { keccak256 } from 'js-sha3';
-import * as SecureStore from 'expo-secure-store'
+import { createWallet } from '../utils/crypto';
+import { saveWallet } from '../utils/storage';
 
 export default function WalletPage() {
     const [publicKey, setPublicKey] = useState<string | undefined>()
     const [privateKey, setPrivateKey] = useState<string | undefined>()
 
     useEffect(() => {
-        createWallet()
+        // generate a wallet
+        const wallet = createWallet()
+
+        setPrivateKey(wallet.private_key)
+        setPublicKey(wallet.public_key)
+
+        // save wallet information to secure storage 
+        saveWallet(wallet.private_key, wallet.public_key, wallet.address as `0x${string}`)
+        .then(() => console.log('wallet generated'))
+        .catch((e) => console.error(e))
+
     }, [])
 
-    function createWallet() {
-        const ec = new EC('secp256k1')
-
-        //Generate a new key pair
-        const keyPair = ec.genKeyPair()
-
-        // get private key
-        const priv = keyPair.getPrivate('hex')
-
-        // Get a public key
-        const pubPoint = keyPair.getPublic()
-        const pub = pubPoint.encode('hex')
-
-        // take keccak hash of public key
-        const pubNoPrefix = pub.slice(2) // dropping the '04' prefix 
-        const keccak = keccak256(Buffer.from(pub, 'hex').toString('base64'))
-
-        // address is 0x + last 20 bytes of public key 
-        let address = keccak.slice(-40)
-
-        // compute checksum for address
-        address = address.toLowerCase().replace('0x', '');
-        const hash = keccak256(address).toString();
-        let result = '0x';
-        for (let i = 0; i < address.length; i++) {
-          if (parseInt(hash[i], 16) >= 8) {
-            result += address[i].toUpperCase();
-          } else {
-            result += address[i];
-          }
-        }
-        const checksum_address = `0x${address}`
-
-        // save wallet info to secure storage
-        saveWallet(priv, pub, checksum_address as `0x${string}`)
-      
-        setPublicKey(pub)
-        setPrivateKey(priv)
-    }
-
-    async function saveWallet(privateKey: string, publicKey: string, address: `0x${string}`) {
-        const wallet_data = {
-            private_key: privateKey,
-            public_ket: publicKey,
-            address: address
-        }
-        await SecureStore.setItemAsync('WALLET', JSON.stringify(wallet_data))
-    }
 
     return (
         <View className="flex-1 justify-between items-center m-10">
