+++ package.json
@@ -15,7 +15,7 @@
     "expo-constants": "~14.4.2",
     "expo-linking": "~5.0.2",
     "expo-router": "^2.0.0",
-    "expo-secure-store": "^12.5.0",
+    "expo-secure-store": "~12.3.1",
     "expo-status-bar": "~1.6.0",
     "js-sha3": "^0.9.2",
     "nativewind": "^2.0.11",
@@ -29,7 +29,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
-    "tailwindcss": "3.3.2" 
+    "tailwindcss": "3.3.2"
   },
   "private": true
 }
